+++ app/core/utils/ext.rb
@@ -2,6 +2,10 @@ module Blank
   def blank?
     respond_to?(:empty?) ? empty? : !self
   end
+
+  def present?
+    !blank?
+  end
 end
 
 module NotNil