+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Fixed
+- Signatures downloading could fail in come cases when the signatures file was larger than 512 megabytes
+
 ## [3.6.0]
 ### Added
 - TurboPatch Support