+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [UNRELEASED]
+## [3.6.1]
 ### Fixed
 - Signatures downloading could fail in come cases when the signatures file was larger than 512 megabytes
 - Diff processing: On Windows, diff processing could end with a permission denied error